feat(table): making a cool table what you wnat
With this module you can create a colored table. The module has
two classes.

* First class `table` has four methods. First methods `colors` it
is used for coloring the table. We don't use it outside class.
Second methods `connect` it is used to connect the top and bottom
parts of the entire table. Method `headers` it is used for the
headers of the tabel, whith this you can corored the text inside
table. Last method `content` has two spaces for text, you can
color the text, each in a different color as you wish.
The inscriptions are separated by a white arrow. We use the
`headers` and `content` methods outside of classes in other
modules to create nice tables.

* Second class `table_elem` is a collection of basic table
elements. You can create and manipulate freely.