feat(emoji): add two class for show entire collection of emoji
This two class `emoji` and `show` are designed so that you can
easily view the entire emoji gallery or just a part of it. The last
lines of code that are comments show how it works, in the next one
I will move these lines to the `tests` folder. I only use this one
`class show` to show all emoticons. The `show` class is based on
the `emoji` class.